mealCalculator with tip choice
Final and best edition of the meal calculator. Default tip amount is 15%. You can change that in the code directly or after you have entered the cost of your meal it'll ask you if you want to change the tip, then it'll give you different option to chose from.
Super edition in the works: after "Is tip amount ok?" input it'll ask if you'd like to insert your own tip amount.